@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -33,22 +33,23 @@ dependencies:
   states_rebuilder: ^3.1.0
 #  meta: ^1.1.8
 #  equatable: ^1.2.0
-  firebase_auth: ^0.20.0+1
-  firebase_core: ^0.7.0
-  cloud_firestore: ^0.16.0
-  charts_flutter: ^0.9.0
-  sqflite: ^1.3.2+3
+  firebase_auth: ^1.0.1
+  firebase_core: ^1.0.1
+  cloud_firestore: ^1.0.1
+  charts_flutter: ^0.10.0
+  sqflite: ^2.0.0+3
   path: ^1.8.0-nullsafety.3
-  intl: ^0.16.1
-  logger: ^0.9.4
-  http: ^0.12.2
-  file_picker: ^2.1.6
+  intl: ^0.17.0
+  logger: ^1.0.0
+  http: ^0.13.0
+  file_picker: ^3.0.0
+  dio: ^4.0.0-prev2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  mockito: ^4.1.4
+  mockito: ^5.0.2
   pedantic: ^1.9.2
 
 
@@ -64,9 +65,9 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+   - assets/certs/ca.crt
+   - assets/certs/smbresthomeclient.p12
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
